projects/mean-reversion/requirements.txt
# Core dependencies
hyperliquid-python-sdk>=0.17.0
python-dotenv>=1.0.0
eth-account>=0.10.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database & caching
supabase>=2.0.0
redis>=5.0.0

# Async & networking
aiohttp>=3.9.0
websockets>=12.0
asyncio-mqtt>=0.16.0

# Monitoring & logging
colorama>=0.4.6
loguru>=0.7.2
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Utilities
pydantic>=2.0.0
python-json-logger>=2.0.7
tenacity>=8.2.0